+++ packages/backend-modules/assets/index.js
@@ -1,4 +1,3 @@
 module.exports = {
   lib: require('./lib'),
-  express: require('./express'),
 }
